Move sanitizer facade/service provider to src.
