+++ src/Facades/Sanitizer/Facade.php
@@ -4,7 +4,7 @@
  * Sanitization functionality to be customized within this project before a 1.0 release.
  */
 
-namespace PerfectOblivion\ActionServiceResponder\Validation\Sanitizer;
+namespace PerfectOblivion\ActionServiceResponder\Facades\Sanitizer;
 
 use Illuminate\Support\Facades\Facade as LaravelFacade;
